hotel-dashboard-backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
pytest>=7.4.0
httpx>=0.25.0
python-dotenv>=1.0.0
alembic>=1.12.0
reportlab>=4.0.0
requests>=2.31.0
python-dateutil>=2.8.2
psycopg2-binary>=2.9.6
